+++ src/Nodes/Node.cpp
@@ -1,7 +1,5 @@
 #include "Node.h"
-#include <stdexcept>
 #include "Cool/String/String.h"
-#include "PrimitiveType.h"
 
 namespace Lab {
 
@@ -12,7 +10,15 @@ auto to_string(Node const& node) -> std::string
 
 auto Node::is_audio_node() const -> bool
 {
-    return Cool::String::contains(_id_names.definition_name, "Audio");
+    return Cool::String::contains(_d.id_names.definition_name, "Audio");
+}
+
+auto Node::as_pod() const -> NodeAsPOD
+{
+    auto pod = NodeAsPOD{_d};
+    for (auto const& input : _value_inputs)
+        pod.value_inputs.push_back(std::visit([](auto&& input) -> Cool::AnyVariable { return input.variable(); }, input));
+    return pod;
 }
 
 } // namespace Lab